refactor and and vue hack
