+++ src/main/java/net/rptools/maptool/client/ui/htmlframe/HTMLWebViewManager.java
@@ -56,6 +56,7 @@ import net.rptools.maptool.client.AppPreferences;
 import net.rptools.maptool.client.MapTool;
 import net.rptools.maptool.client.functions.MacroLinkFunction;
 import net.rptools.maptool.client.functions.json.JSONMacroFunctions;
+import net.rptools.maptool.client.ui.htmlframe.content.HTMLContent;
 import net.rptools.maptool.language.I18N;
 import net.rptools.maptool.model.TextMessage;
 import net.rptools.maptool.model.library.LibraryManager;
@@ -334,7 +335,8 @@ public class HTMLWebViewManager {
     if (htmlContent.isUrl()) {
       webEngine.load(htmlContent.getUrl().toString());
     } else {
-      webEngine.load(htmlContent.injectJavaBridge().getHtmlStringAsDataUrl());
+      webEngine.load(
+          htmlContent.injectJSAPI().InjectContentSecurityPolicy().getHtmlStringAsDataUrl());
     }
   }
